pyproject.toml
[project]
name = "silaba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "gTTS>=2.5.4",
    "google-cloud-texttospeech>=2.26.0",
    "httpx>=0.27.2",
    "pyglet>=2.0.17",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
